Guard Costumes
So I added how the guard will look whether it's facing up, down, right,
or left. My next step is when the guard walks into the wall, have it
change costumes to make the guard face the opposite direction and walk
that way, until it reaches the wall again, and repeat.